cpufreq: interactive: teach interactive to be interactive
This squashes in a few commits for interactive from this branch;
https://github.com/franciscofranco/hammerhead/commit/db6c92d0ede1d09eaaf143e56ea858d7e398384a

At this point we can use the generic touchboost implementation
in the kernel to boost frequencies when touch events are happening.

This makes the library in /system/lib/hw/ obsolete and power.msm8226.so
should be deleted. touchboost contains a fix for a missing error handling.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>

Conflicts:
	drivers/cpufreq/cpufreq_interactive.c

Conflicts:
	drivers/input/touchscreen/Makefile